refactor: 데이터테이블 스토리북 파일 이동(#129)
- 데이터테이블 스토리북 파일 이동
