+++ src/components/common/common-table/DataTableWithPagination.stories.tsx
@@ -6,9 +6,9 @@ import {
   useReactTable,
 } from '@tanstack/react-table'
 
-import { DataTable } from '../DataTable'
-import { TablePagination } from '../TablePagination'
+import { DataTable } from './DataTable'
 import { mockUsers, userColumns } from './table-mock-data'
+import { TablePagination } from './TablePagination'
 
 const meta: Meta = {
   title: 'Components/DataTable/WithPagination',
